288884 - Updated API Endpoints for CustomApplicationStatus, Moved the logic to Upsert into a single command, Updated ApplicationStatus as an Enum and made the Label nullable for the entity
